remove unused range class for raw-orche
