+++ sharding-jdbc-example/raw-jdbc-orche-example/src/main/java/io/shardingsphere/example/jdbc/orche/factory/OrchestrationDataSourceFactory.java
@@ -25,10 +25,10 @@ import io.shardingsphere.example.jdbc.orche.config.cloud.CloudShardingDatabasesC
 import io.shardingsphere.example.jdbc.orche.config.cloud.CloudShardingMasterSlaveConfiguration;
 import io.shardingsphere.example.jdbc.orche.config.cloud.CloudShardingTablesConfiguration;
 import io.shardingsphere.example.jdbc.orche.config.local.LocalMasterSlaveConfiguration;
-import io.shardingsphere.example.jdbc.orche.config.local.LocalShardingDatabasesAndTablesConfigurationPrecise;
-import io.shardingsphere.example.jdbc.orche.config.local.LocalShardingDatabasesConfigurationPrecise;
-import io.shardingsphere.example.jdbc.orche.config.local.LocalShardingMasterSlaveConfigurationPrecise;
-import io.shardingsphere.example.jdbc.orche.config.local.LocalShardingTablesConfigurationPrecise;
+import io.shardingsphere.example.jdbc.orche.config.local.LocalShardingDatabasesAndTablesConfiguration;
+import io.shardingsphere.example.jdbc.orche.config.local.LocalShardingDatabasesConfiguration;
+import io.shardingsphere.example.jdbc.orche.config.local.LocalShardingMasterSlaveConfiguration;
+import io.shardingsphere.example.jdbc.orche.config.local.LocalShardingTablesConfiguration;
 import io.shardingsphere.example.type.RegistryCenterType;
 import io.shardingsphere.example.type.ShardingType;
 import org.apache.shardingsphere.orchestration.reg.api.RegistryCenterConfiguration;
@@ -43,19 +43,19 @@ public class OrchestrationDataSourceFactory {
         ExampleConfiguration configuration;
         switch (shardingType) {
             case SHARDING_DATABASES:
-                configuration = loadConfigFromRegCenter ? new CloudShardingDatabasesConfiguration(registryCenterConfig) : new LocalShardingDatabasesConfigurationPrecise(registryCenterConfig);
+                configuration = loadConfigFromRegCenter ? new CloudShardingDatabasesConfiguration(registryCenterConfig) : new LocalShardingDatabasesConfiguration(registryCenterConfig);
                 break;
             case SHARDING_TABLES:
-                configuration = loadConfigFromRegCenter ? new CloudShardingTablesConfiguration(registryCenterConfig) : new LocalShardingTablesConfigurationPrecise(registryCenterConfig);
+                configuration = loadConfigFromRegCenter ? new CloudShardingTablesConfiguration(registryCenterConfig) : new LocalShardingTablesConfiguration(registryCenterConfig);
                 break;
             case SHARDING_DATABASES_AND_TABLES:
-                configuration = loadConfigFromRegCenter ? new CloudShardingDatabasesAndTablesConfiguration(registryCenterConfig) : new LocalShardingDatabasesAndTablesConfigurationPrecise(registryCenterConfig);
+                configuration = loadConfigFromRegCenter ? new CloudShardingDatabasesAndTablesConfiguration(registryCenterConfig) : new LocalShardingDatabasesAndTablesConfiguration(registryCenterConfig);
                 break;
             case MASTER_SLAVE:
                 configuration = loadConfigFromRegCenter ? new CloudMasterSlaveConfiguration(registryCenterConfig) : new LocalMasterSlaveConfiguration(registryCenterConfig);
                 break;
             case SHARDING_MASTER_SLAVE:
-                configuration = loadConfigFromRegCenter ? new CloudShardingMasterSlaveConfiguration(registryCenterConfig) : new LocalShardingMasterSlaveConfigurationPrecise(registryCenterConfig);
+                configuration = loadConfigFromRegCenter ? new CloudShardingMasterSlaveConfiguration(registryCenterConfig) : new LocalShardingMasterSlaveConfiguration(registryCenterConfig);
                 break;
             default:
                 throw new UnsupportedOperationException(shardingType.name());
